+++ src/main/java/net/maritimecloud/identityregistry/services/VesselService.java
@@ -27,5 +27,5 @@ public interface VesselService {
 
     Iterable<Vessel> listAllVessels();
 
-    List<Vessel> listOrgVessels(int id);
+    List<Vessel> listOrgVessels(Long id);
 }
